+++ app/models/group_lesson_summary.rb
@@ -17,7 +17,7 @@ class GroupLessonSummary < ApplicationRecord
   def self.select_for_graphing(relation, sorting_direction, number_of_elements)
     relation
       .order(lesson_date: sorting_direction).limit(number_of_elements)
-      .select(Arel.sql('extract(EPOCH FROM lesson_date) :: INT AS timestamp'), :average_mark, Arel.sql("to_char(lesson_date, 'MM/DD/YY') AS date"), :lesson_uid, :lesson_id)
+      .select(Arel.sql('extract(EPOCH FROM lesson_date) :: INT AS timestamp'), :average_mark, :lesson_date, :lesson_uid, :lesson_id)
   end
 
   def self.process_around_result(lesson, result, number_of_elements)
@@ -35,7 +35,7 @@ class GroupLessonSummary < ApplicationRecord
   def self.find_closest_index(lessons, lesson)
     previous_diff = nil
     lessons.each_with_index.reduce(nil) do |previous_closest_i, (l, i)|
-      diff = (lesson.date.to_time.to_i - l.date.to_time.to_i).abs
+      diff = (lesson.date.to_time.to_i - l.lesson_date.to_time.to_i).abs
       if previous_diff.nil? || diff < previous_diff
         previous_diff = diff
         return i