+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Not displaying deleted skills when creating subjects
 - Implemented student ordering in single lesson view 
 - Default Lessons view order
+- Added color indicators for student grading status in single lesson view
 
 ## 0.21.3 - Logs
 - Updated Logging aggregation