+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Default Lessons view order
 - Added color indicators for student grading status in single lesson view
 - Added a display of graded and total students in group in the lesson table
+- Fix Lesson Graph not showing on the first load
+- Added Average Grade column to lesson table
 
 ## 0.21.3 - Logs
 - Updated Logging aggregation